chore(web): lazy-load dashboard
